abstract out myApp to its own class
