+++ lib/main_prod.dart
@@ -1,3 +1,4 @@
+import 'package:flap_app/my_app.dart';
 import 'package:flap_app/presentation/navigation.dart';
 import 'package:flap_app/presentation/theme/app_theme.dart';
 import 'package:flutter/material.dart';
@@ -10,24 +11,4 @@ void main() {
     "baseUrl": "api.spoonacular.com" //replace this baseUrl with the baseUrl used for your production API
   });
   runApp(const ProviderScope(child: MyApp()));
-}
-
-class MyApp extends StatelessWidget {
-  const MyApp({super.key});
-
-  // This widget is the root of your application.
-  @override
-  Widget build(BuildContext context) {
-    const theme = AppTheme();
-    // ProviderScope is what makes Riverpod work.
-    // MaterialApp.router is what sets our routerConfig to our GoRouter object
-    return FlavorBanner(
-      child: MaterialApp(
-        title: FlavorConfig.instance.variables["appTitle"],
-        theme: theme.toThemeData(),
-        darkTheme: theme.toThemeDataDark(),
-        home: MaterialApp.router(routerConfig: navigationRouter),
-      ),
-    );
-  }
 }
